Abstract test case class moved to common directory
