+++ tests/bootstrap.php
@@ -1,6 +1,6 @@
 <?php
 
 require __DIR__ . '/../vendor/autoload.php';
-require __DIR__ . '/Keboola/Test/ExtractorTestCase.php';
+require __DIR__ . '/Keboola/MongoDbExtractor/ExtractorTestCase.php';
 
 define('ROOT_PATH', __DIR__ . '/..');
